requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
certifi==2026.2.25
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
gast==0.7.0
google-pasta==0.2.0
googleapis-common-protos==1.74.0
grandalf==0.8
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
kubernetes==35.0.0
langchain==1.2.15
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-ollama==1.0.1
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2

opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
pandas==2.3.3
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pypdf==6.9.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
torch
torch-vision
tqdm==4.67.3
transformers 
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wikipedia==1.4.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
youtube-transcript-api
tiktoken 
